backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
opencv-python-headless==4.10.0.84
numpy==1.26.4
websockets==12.0
pillow==10.4.0
depthai
sqlalchemy>=2.0.0
alembic>=1.13.0
asyncpg>=0.29.0
web3>=6.0.0
eth-account>=0.10.0
python-dotenv>=1.0.0
